+++ src/editors/query/query.uql.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { InlineFormLabel, CodeEditor, CodeEditorSuggestionItem, Icon, CodeEditorSuggestionItemKind } from '@grafana/ui';
 import { InfinityQuery, EditorMode } from '../../types';
+declare const monaco: any;
 
 export const UQLEditor = (props: { mode: EditorMode; query: InfinityQuery; onChange: (value: InfinityQuery) => void; onRunQuery: () => void }) => {
   const { query, mode, onChange, onRunQuery } = props;
@@ -11,10 +12,11 @@ export const UQLEditor = (props: { mode: EditorMode; query: InfinityQuery; onCha
       onRunQuery();
     }
   };
+  const handleMount = (editor: any) => registerUQL(editor);
   const getUQLSuggestions = (): CodeEditorSuggestionItem[] => {
     return [
-      { label: 'project', kind: CodeEditorSuggestionItemKind.Method },
-      { label: 'project-away', kind: CodeEditorSuggestionItemKind.Method },
+      ...UQLKeyWords.map((item: string) => ({ label: item, kind: CodeEditorSuggestionItemKind.Method })),
+      ...UQLFunctions.map((item: string) => ({ label: item, kind: CodeEditorSuggestionItemKind.Method })),
     ];
   };
   return query.type === 'uql' ? (
@@ -33,6 +35,7 @@ export const UQLEditor = (props: { mode: EditorMode; query: InfinityQuery; onCha
           getSuggestions={getUQLSuggestions}
           onSave={onUQLChange}
           onBlur={onUQLChange}
+          onEditorDidMount={handleMount}
         />
         <Icon name="play" size="lg" style={{ color: 'greenyellow' }} onClick={() => {}} />
       </div>
@@ -41,3 +44,56 @@ export const UQLEditor = (props: { mode: EditorMode; query: InfinityQuery; onCha
     <></>
   );
 };
+
+async function registerUQL(editor: any) {
+  editor.updateOptions({ fixedOverflowWidgets: true });
+  monaco.languages.register({ id: 'uql' });
+}
+
+const UQLKeyWords = ['parse-json', 'parse-csv', 'parse-xml', 'parse-yaml', 'project', 'project-away', 'extend', 'scope', 'summarize', 'mv-expand', 'order by'];
+const UQLFunctions = [
+  'count',
+  'sum',
+  'diff',
+  'mul',
+  'div',
+  'min',
+  'max',
+  'mean',
+  'first',
+  'last',
+  'latest',
+  'strcat',
+  'dcount',
+  'distinct',
+  'random',
+  'toupper',
+  'tolower',
+  'strlen',
+  'trim',
+  'trim_start',
+  'trim_end',
+  'toint',
+  'tolong',
+  'tonumber',
+  'tobool',
+  'tostring',
+  'todouble',
+  'tofloat',
+  'parse_url',
+  'parse_urlquery',
+  'todatetime',
+  'tounixtime',
+  'unixtime_seconds_todatetime',
+  'unixtime_nanoseconds_todatetime',
+  'unixtime_milliseconds_todatetime',
+  'unixtime_microseconds_todatetime',
+  'format_datetime',
+  'add_datetime',
+  'startofminute',
+  'startofhour',
+  'startofday',
+  'startofmonth',
+  'startofweek',
+  'startofyear',
+];